Issue 230_1: Remove ConfigDefs
