+++ app/src/main/java/org/apache/roller/weblogger/business/themes/ThemeManagerImpl.java
@@ -454,26 +454,4 @@ public class ThemeManagerImpl implements ThemeManager {
         return new String(chars, 0, length);
     }
 
-	/**
-	 * @see ThemeManager#reLoadThemeFromDisk(String)
-	 */
-	public boolean reLoadThemeFromDisk(String reloadTheme) {
-		boolean reloaded = false;
-
-		try {
-            SharedTheme theme = loadThemeData(reloadTheme);
-            Theme loadedTheme = themes.get(theme.getId());
-
-            if (loadedTheme != null && theme.getLastModified().after(
-                    loadedTheme.getLastModified())) {
-                themes.remove(theme.getId());
-                themes.put(theme.getId(), theme);
-                reloaded = true;
-            }
-		} catch (Exception unexpected) {
-			// shouldn't happen, so let's learn why it did
-			log.error("Problem reloading theme " + reloadTheme, unexpected);
-		}
-		return reloaded;
-	}
 }
