+++ app/src/main/java/org/apache/roller/weblogger/ui/rendering/servlets/PageServlet.java
@@ -28,14 +28,12 @@ import org.apache.roller.weblogger.WebloggerException;
 import org.apache.roller.weblogger.business.HitCountQueue;
 import org.apache.roller.weblogger.business.WeblogEntryManager;
 import org.apache.roller.weblogger.business.WebloggerFactory;
-import org.apache.roller.weblogger.business.themes.ThemeManager;
 import org.apache.roller.weblogger.config.WebloggerConfig;
 import org.apache.roller.weblogger.config.WebloggerRuntimeConfig;
 import org.apache.roller.weblogger.pojos.ThemeTemplate;
 import org.apache.roller.weblogger.pojos.ThemeTemplate.ComponentType;
 import org.apache.roller.weblogger.pojos.Weblog;
 import org.apache.roller.weblogger.pojos.WeblogEntry;
-import org.apache.roller.weblogger.pojos.WeblogTheme;
 import org.apache.roller.weblogger.ui.core.RollerContext;
 import org.apache.roller.weblogger.ui.rendering.Renderer;
 import org.apache.roller.weblogger.ui.rendering.RendererManager;
@@ -47,7 +45,6 @@ import org.apache.roller.weblogger.ui.rendering.util.WeblogPageRequest;
 import org.apache.roller.weblogger.ui.rendering.util.cache.SiteWideCache;
 import org.apache.roller.weblogger.ui.rendering.util.cache.WeblogPageCache;
 import org.apache.roller.weblogger.util.Blacklist;
-import org.apache.roller.weblogger.util.I18nMessages;
 import org.apache.roller.weblogger.util.cache.CachedContent;
 
 import javax.servlet.ServletConfig;
@@ -87,9 +84,6 @@ public class PageServlet extends HttpServlet {
     private boolean excludeOwnerPages = false;
     private WeblogPageCache weblogPageCache = null;
     private SiteWideCache siteWideCache = null;
-
-    // Development theme reloading
-    Boolean themeReload = false;
 
     /**
      * Init method for this servlet
@@ -131,9 +125,6 @@ public class PageServlet extends HttpServlet {
                                 + "'.  Robots will not be filtered. ", e);
             }
         }
-
-        // Development theme reloading
-        themeReload = WebloggerConfig.getBooleanProperty("themes.reload.mode");
     }
 
     /**
@@ -211,26 +202,6 @@ public class PageServlet extends HttpServlet {
             cacheKey = siteWideCache.generateKey(pageRequest);
         } else {
             cacheKey = weblogPageCache.generateKey(pageRequest);
-        }
-
-        // Development only. Reload if theme has been modified
-        if (themeReload && !weblog.getEditorTheme().equals(WeblogTheme.CUSTOM)
-                && (pageRequest.getPathInfo() == null || !pageRequest.getPathInfo().endsWith(".css"))) {
-            try {
-                ThemeManager manager = WebloggerFactory.getWeblogger().getThemeManager();
-                boolean reloaded = manager.reLoadThemeFromDisk(weblog.getEditorTheme());
-                if (reloaded) {
-                    if (isSiteWide) {
-                        siteWideCache.clear();
-                    } else {
-                        weblogPageCache.clear();
-                    }
-                    I18nMessages.reloadBundle(weblog.getLocaleInstance());
-                }
-
-            } catch (Exception ex) {
-                log.error("ERROR - reloading theme " + ex);
-            }
         }
 
         // cached content checking
