+++ app/src/main/resources/org/apache/roller/weblogger/config/roller.properties
@@ -123,13 +123,6 @@ search.index.dir=${user.home}/roller_data/search-index
 # The directory in which Roller will look for themes
 themes.dir=${webapp.context}
 
-# Allow theme reloading during development
-# 1. You must be logged in.
-# 2. Delete the style-sheet (Design | Style Sheet tab) from the weblog/theme you're changing.
-# 3. To reload macros under path /WEB-INF/velocity, first modify the macro file and then trigger 
-#    the reload by changing a template on the theme currently working on, then refresh the page.
-themes.reload.mode=false
-
 #-----------------------------------------------------------------------------
 # Feature specific settings
 #-----------------------------------------------------------------------------
