+++ app/src/test/resources/roller-jettyrun.properties
@@ -32,9 +32,6 @@ log4j.appender.roller.File=${project.build.directory}/work/roller.log
 # don't auto migrate during tests
 uploads.migrate.auto=false
 
-# turn on theme reloading
-themes.reload.mode=true
-
 # turn off caching for testing/debugging
 cache.sitewide.enabled=false
 cache.weblogpage.enabled=false
