+++ it-selenium/src/test/resources/roller-jettyrun.properties
@@ -30,9 +30,6 @@ search.index.dir          =target/work/search-index
 mediafiles.storage.dir    =target/work/mediafiles
 log4j.appender.roller.File=target/work/roller.log
 
-# turn on theme reloading
-themes.reload.mode=true
-
 # turn off caching for testing/debugging
 cache.sitewide.enabled=false
 cache.weblogpage.enabled=false
